+++ tests/conftest.py
@@ -3,9 +3,12 @@
 import pytest
 
 from cool_seq_tool import CoolSeqTool
-from cool_seq_tool.mappers.exon_genomic_coords import _ExonCoord
 from cool_seq_tool.schemas import ManeGeneData, Strand
-from cool_seq_tool.sources.uta_database import GenomicAlnData, GenomicTxMetadata
+from cool_seq_tool.sources.uta_database import (
+    ExonCoord,
+    GenomicAlnData,
+    GenomicTxMetadata,
+)
 
 
 @pytest.fixture(scope="session")
@@ -42,7 +45,7 @@ def test_mane_transcript_mappings(test_cool_seq_tool):
 def nm_152263_exons():
     """Create test fixture for NM_152263.3 exons."""
     return [
-        _ExonCoord(
+        ExonCoord(
             ord=0,
             tx_start_i=0,
             tx_end_i=234,
@@ -50,7 +53,7 @@ def nm_152263_exons():
             alt_end_i=154192135,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=1,
             tx_start_i=234,
             tx_end_i=360,
@@ -58,7 +61,7 @@ def nm_152263_exons():
             alt_end_i=154191311,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=2,
             tx_start_i=360,
             tx_end_i=494,
@@ -66,7 +69,7 @@ def nm_152263_exons():
             alt_end_i=154176248,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=3,
             tx_start_i=494,
             tx_end_i=612,
@@ -74,7 +77,7 @@ def nm_152263_exons():
             alt_end_i=154173201,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=4,
             tx_start_i=612,
             tx_end_i=683,
@@ -82,7 +85,7 @@ def nm_152263_exons():
             alt_end_i=154172978,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=5,
             tx_start_i=683,
             tx_end_i=759,
@@ -90,7 +93,7 @@ def nm_152263_exons():
             alt_end_i=154171488,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=6,
             tx_start_i=759,
             tx_end_i=822,
@@ -98,7 +101,7 @@ def nm_152263_exons():
             alt_end_i=154170711,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=7,
             tx_start_i=822,
             tx_end_i=892,
@@ -106,7 +109,7 @@ def nm_152263_exons():
             alt_end_i=154170469,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=8,
             tx_start_i=892,
             tx_end_i=971,
@@ -114,7 +117,7 @@ def nm_152263_exons():
             alt_end_i=154169383,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=9,
             tx_start_i=971,
             tx_end_i=7099,
@@ -129,7 +132,7 @@ def nm_152263_exons():
 def nm_152263_exons_genomic_coords():
     """Create test fixture for NM_152263.4 exons and genomic coordinates."""
     return [
-        _ExonCoord(
+        ExonCoord(
             ord=0,
             tx_start_i=0,
             tx_end_i=199,
@@ -137,7 +140,7 @@ def nm_152263_exons_genomic_coords():
             alt_end_i=154192100,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=1,
             tx_start_i=199,
             tx_end_i=325,
@@ -145,7 +148,7 @@ def nm_152263_exons_genomic_coords():
             alt_end_i=154191311,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=2,
             tx_start_i=325,
             tx_end_i=459,
@@ -153,7 +156,7 @@ def nm_152263_exons_genomic_coords():
             alt_end_i=154176248,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=3,
             tx_start_i=459,
             tx_end_i=577,
@@ -161,7 +164,7 @@ def nm_152263_exons_genomic_coords():
             alt_end_i=154173201,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=4,
             tx_start_i=577,
             tx_end_i=648,
@@ -169,7 +172,7 @@ def nm_152263_exons_genomic_coords():
             alt_end_i=154172978,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=5,
             tx_start_i=648,
             tx_end_i=724,
@@ -177,7 +180,7 @@ def nm_152263_exons_genomic_coords():
             alt_end_i=154171488,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=6,
             tx_start_i=724,
             tx_end_i=787,
@@ -185,7 +188,7 @@ def nm_152263_exons_genomic_coords():
             alt_end_i=154170711,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=7,
             tx_start_i=787,
             tx_end_i=857,
@@ -193,7 +196,7 @@ def nm_152263_exons_genomic_coords():
             alt_end_i=154170469,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=8,
             tx_start_i=857,
             tx_end_i=936,
@@ -201,7 +204,7 @@ def nm_152263_exons_genomic_coords():
             alt_end_i=154169383,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=9,
             tx_start_i=936,
             tx_end_i=7064,
@@ -216,7 +219,7 @@ def nm_152263_exons_genomic_coords():
 def nm_001105539_exons_genomic_coords():
     """Create test fixture for NM_001105539.3 exons and genomic coordinates."""
     return [
-        _ExonCoord(
+        ExonCoord(
             ord=0,
             tx_start_i=0,
             tx_end_i=1557,
@@ -224,7 +227,7 @@ def nm_001105539_exons_genomic_coords():
             alt_end_i=80487782,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=1,
             tx_start_i=1557,
             tx_end_i=2446,
@@ -232,7 +235,7 @@ def nm_001105539_exons_genomic_coords():
             alt_end_i=80500382,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=2,
             tx_start_i=2446,
             tx_end_i=2545,
@@ -240,7 +243,7 @@ def nm_001105539_exons_genomic_coords():
             alt_end_i=80514008,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=3,
             tx_start_i=2545,
             tx_end_i=2722,
@@ -248,7 +251,7 @@ def nm_001105539_exons_genomic_coords():
             alt_end_i=80518579,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=4,
             tx_start_i=2722,
             tx_end_i=2895,
@@ -256,7 +259,7 @@ def nm_001105539_exons_genomic_coords():
             alt_end_i=80518954,
             alt_strand=Strand.NEGATIVE,
         ),
-        _ExonCoord(
+        ExonCoord(
             ord=5,
             tx_start_i=2895,
             tx_end_i=9938,
@@ -271,7 +274,7 @@ def nm_001105539_exons_genomic_coords():
 def mm_001005183_1_exons():
     """Create test fixture for NM_001005183.1 exons and genomic coordinates"""
     return [
-        _ExonCoord(
+        ExonCoord(
             ord=0,
             tx_start_i=0,
             tx_end_i=939,
